add post result function

+++ src/services/apiAdmin.js
@@ -3,7 +3,7 @@ import handleApiError from '../utils/handleApiError';
 
 export async function getAdminThemes() {
   try {
-    const response = await api.get('http://localhost:3000/api/v1/admin/themes');
+    const response = await api.get('/api/v1/admin/themes');
     // console.log(response.data);
     return response.data; // Return the data if needed
   } catch (error) {
@@ -15,10 +15,7 @@ export async function getAdminThemes() {
 
 export async function postTheme(data) {
   try {
-    const response = await api.post(
-      'http://localhost:3000/api/v1/admin/themes',
-      data,
-    );
+    const response = await api.post('/api/v1/admin/themes', data);
     // console.log(response.data);
     return response.data; // Return the data if needed
   } catch (error) {
@@ -31,10 +28,7 @@ export async function postTheme(data) {
 export async function updateTheme(data, id) {
   console.log('updateTheme', data, id);
   try {
-    const response = await api.patch(
-      `http://localhost:3000/api/v1/admin/themes/${id}`,
-      data,
-    );
+    const response = await api.patch(`/api/v1/admin/themes/${id}`, data);
     // console.log(response.data);
     return response.data; // Return the data if needed
   } catch (error) {
@@ -45,10 +39,33 @@ export async function updateTheme(data, id) {
 }
 
 export async function deleteTheme(id) {
+  console.log('deleteTheme', id);
   try {
-    const response = await api.delete(
-      `http://localhost:3000/api/v1/admin/themes/${id}`,
-    );
+    const response = await api.delete(`/api/v1/admin/themes/${id}`);
+    // console.log(response.data);
+    return response.data; // Return the data if needed
+  } catch (error) {
+    console.error('Error fetching data:', error); // Handle error
+    handleApiError(error);
+    throw error;
+  }
+}
+
+export async function createPrepResult(data) {
+  try {
+    const response = await api.post('/api/v1/results', data);
+    // console.log(response.data);
+    return response.data; // Return the data if needed
+  } catch (error) {
+    console.error('Error fetching data:', error); // Handle error
+    handleApiError(error);
+    throw error;
+  }
+}
+
+export async function getAdminResults() {
+  try {
+    const response = await api.get('/api/v1/results');
     // console.log(response.data);
     return response.data; // Return the data if needed
   } catch (error) {
